django_site/requirements.txt
astroid==1.6.1
autopep8==1.3.4
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.9
dj-database-url==0.5.0
Django==2.0.2
django-bootstrap3==9.1.0
django-bootstrap4==0.0.6
django-heroku==0.3.1
djangorestframework==3.7.7
gunicorn==19.7.1
idna==2.6
isort==4.3.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.0.0
psutil==5.4.3
psycopg2==2.7.4
pycodestyle==2.3.1
pylint==1.8.2
pytz==2018.3
requests==2.18.4
six==1.11.0
urllib3==1.22
whitenoise==3.3.1
wrapt==1.10.11
